requirements.txt
fastapi
uvicorn[standard]
pydantic
typer
r2pipe
requests
fastmcp
pefile
pyelftools
python-dotenv
pytest
